backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.16.1
opencv-python==4.10.0.84
numpy>=1.26.0
Pillow==10.1.0
pydantic==2.5.0